# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduAgent API",
    "description": "EduAgent API with Azure Entra Authentication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/attempts": {
      "post": {
        "tags": ["attempts"],
        "summary": "Create an attempt record",
        "description": "Create a new study attempt record for a flashcard or quiz question",
        "operationId": "create_attempt_v1_attempts_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Project Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAttemptRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AttemptResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["attempts"],
        "summary": "List attempt records",
        "description": "List study attempt records for the current user, optionally filtered by project",
        "operationId": "list_attempts_v1_attempts_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Optional project ID filter",
              "title": "Project Id"
            },
            "description": "Optional project ID filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AttemptListResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/attempts/batch": {
      "post": {
        "tags": ["attempts"],
        "summary": "Create multiple attempt records",
        "description": "Create multiple study attempt records in a single batch operation",
        "operationId": "create_attempts_batch_v1_attempts_batch_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Project Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttemptBatchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AttemptListResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptDto": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "user_id": { "type": "string", "title": "User Id" },
          "project_id": { "type": "string", "title": "Project Id" },
          "item_type": { "type": "string", "title": "Item Type" },
          "item_id": { "type": "string", "title": "Item Id" },
          "topic": { "type": "string", "title": "Topic" },
          "user_answer": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Answer"
          },
          "correct_answer": { "type": "string", "title": "Correct Answer" },
          "was_correct": { "type": "boolean", "title": "Was Correct" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "project_id",
          "item_type",
          "item_id",
          "topic",
          "correct_answer",
          "was_correct",
          "created_at"
        ],
        "title": "AttemptDto",
        "description": "Attempt data transfer object."
      },
      "AttemptListResponse": {
        "properties": {
          "attempts": {
            "items": { "$ref": "#/components/schemas/AttemptDto" },
            "type": "array",
            "title": "Attempts"
          },
          "total": { "type": "integer", "title": "Total" }
        },
        "type": "object",
        "required": ["attempts", "total"],
        "title": "AttemptListResponse",
        "description": "Response model for listing attempts."
      },
      "AttemptResponse": {
        "properties": {
          "attempt": { "$ref": "#/components/schemas/AttemptDto" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["attempt", "message"],
        "title": "AttemptResponse",
        "description": "Response model for attempt operations."
      },
...
      "CreateAttemptBatchRequest": {
        "properties": {
          "attempts": {
            "items": { "$ref": "#/components/schemas/CreateAttemptRequest" },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Attempts",
            "description": "List of attempts to create"
          }
        },
        "type": "object",
        "required": ["attempts"],
        "title": "CreateAttemptBatchRequest",
        "description": "Request model for creating multiple attempt records."
      },
      "CreateAttemptRequest": {
        "properties": {
          "item_type": {
            "type": "string",
            "pattern": "^(flashcard|quiz)$",
            "title": "Item Type",
            "description": "Type of item: flashcard or quiz"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id",
            "description": "ID of the flashcard or quiz question"
          },
          "topic": {
            "type": "string",
            "maxLength": 500,
            "title": "Topic",
            "description": "Topic extracted from question"
          },
          "user_answer": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Answer",
            "description": "User's answer (only for quizzes, null for flashcards)"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer",
            "description": "The correct answer - flashcard answer or quiz correct option"
          },
          "was_correct": {
            "type": "boolean",
            "title": "Was Correct",
            "description": "Whether the user got it right"
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "item_id",
          "topic",
          "correct_answer",
          "was_correct"
        ],
        "title": "CreateAttemptRequest",
        "description": "Request model for creating an attempt record."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "AzureEntra": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com/ce716f69-d4fc-4efe-999b-47553cdcbc3d/oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com/ce716f69-d4fc-4efe-999b-47553cdcbc3d/oauth2/v2.0/token",
            "scopes": {
              "api://04598a73-4879-4c7b-9115-fef83c59fba6/user_impersonation": "Access API as user",
              "openid": "OpenID Connect",
              "profile": "User profile",
              "email": "User email"
            }
          },
          "implicit": {
            "authorizationUrl": "https://login.microsoftonline.com/ce716f69-d4fc-4efe-999b-47553cdcbc3d/oauth2/v2.0/authorize",
            "scopes": {
              "openid": "OpenID Connect",
              "profile": "User profile",
              "email": "User email"
            }
          }
        }
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "BearerAuth": [] }]
}